Trims down the logic of the isCheck function.
The right hand side of the || operator contained both a check for item.data.date.proficient, and a check for the type of this value. If the value is not defined, then the typeof operator does the work of both of these checks, because the type will be undefined, not number.

Since the right operand for || can no longer yield undefined, the conditional expression is no longer necessary, as both operands yield a Boolean directly.

As the long property chain is no longer being repeated, we can also throw out the itemData variable declaration, as well as the one for output, returning the result directly.